+++ BE/src/main/java/udpm/hn/server/test/core/staff/processor/service/impl/ProcessorServiceImpl.java
@@ -7,6 +7,8 @@ import udpm.hn.server.test.core.staff.processor.repository.ADProcessorRepository
 import udpm.hn.server.test.core.staff.processor.service.ProcessorService;
 import udpm.hn.server.test.core.common.base.ResponseObject;
 import udpm.hn.server.test.entity.Processor;
+import udpm.hn.server.test.repository.ProcessorRepository;
+import udpm.hn.server.test.repository.ProductRepository;
 
 import java.util.Optional;
 
@@ -14,6 +16,8 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ProcessorServiceImpl implements ProcessorService {
     private final ADProcessorRepository techSpecRepository;
+
+    private final ProductRepository productRepository;
 
     @Override
     public ResponseObject<?> getAllTechSpec() {
@@ -31,10 +35,10 @@ public class ProcessorServiceImpl implements ProcessorService {
         if (optionalProcessor.isPresent()) {
             return new ResponseObject<>(null , HttpStatus.CONFLICT , "Đã tồn tại Processor");
         }
-          processor.setProcessorName(processor.getProcessorName());
-          Processor add = techSpecRepository.save(processor);
+        processor.setProcessorName(processor.getProcessorName());
+         techSpecRepository.save(processor);
 
-        return new ResponseObject<>(add , HttpStatus.ACCEPTED , "Thêm thành cong");
+        return new ResponseObject<>(null , HttpStatus.ACCEPTED , "Thêm thành cong");
     }
 
     @Override
@@ -45,14 +49,19 @@ public class ProcessorServiceImpl implements ProcessorService {
         }
 
         Optional<Processor> optionalProcessorName = techSpecRepository.findByProcessorName(processor.getProcessorName());
-        if (optionalProcessorName.isPresent() && !optionalProcessor.get().getId().equals(id)) {
+        if (optionalProcessorName.isPresent() && !optionalProcessorName.get().getId().equals(id)) {
             return new ResponseObject<>(null, HttpStatus.CONFLICT, "Đã tồn tại Processor");
         }
+
         Processor processorToUpdate = optionalProcessor.get();
+
+        processorToUpdate.setProcessorName(processor.getProcessorName());
+
         techSpecRepository.save(processorToUpdate);
 
         return new ResponseObject<>(processorToUpdate, HttpStatus.ACCEPTED, "Cập nhật thành công");
     }
+
 
     @Override
     public ResponseObject<?> deleteTechSpec(String id) {
@@ -61,8 +70,22 @@ public class ProcessorServiceImpl implements ProcessorService {
             return new ResponseObject<>(null , HttpStatus.NOT_FOUND , "không tìm thấy id này");
         }
 
+        boolean isProcessor = productRepository.existsByProcessorId(id);
+        if (isProcessor) {
+            return new ResponseObject<>(null, HttpStatus.BAD_REQUEST,
+                    "Không thể xoá processor này vì đang được sử dụng trong sản phẩm!");
+        }
+
         techSpecRepository.deleteById(id);
         return new ResponseObject<>(null , HttpStatus.ACCEPTED , "xóa thành công");
     }
 
+    @Override
+    public ResponseObject<?> findProcessorID(String id){
+        return new ResponseObject<>(
+                techSpecRepository.findByIdProcessor(id),
+                HttpStatus.OK,
+                "Lây thành công processor"
+        );
+    }
 }
