+++ BE/src/main/java/udpm/hn/server/test/core/staff/version/service/impl/VersionServiceImpl.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import udpm.hn.server.test.core.common.base.ResponseObject;
 import udpm.hn.server.test.core.staff.version.service.VersionService;
 import udpm.hn.server.test.entity.Version;
+import udpm.hn.server.test.repository.ProductDetailRepository;
+import udpm.hn.server.test.repository.ProductRepository;
 import udpm.hn.server.test.repository.VersionRepository;
 
 import java.util.Optional;
@@ -14,6 +16,8 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class VersionServiceImpl implements VersionService {
     private final VersionRepository versionRepository;
+
+    private final ProductDetailRepository productDetailRepository;
 
     @Override
     public ResponseObject<?> getAllVersion(){
@@ -57,7 +61,19 @@ public class VersionServiceImpl implements VersionService {
             return new ResponseObject<>(null, HttpStatus.NOT_FOUND, "khong ton tai Version");
         }
 
+        boolean isVersion = productDetailRepository.existsByVersionId(id);
+        if(isVersion){
+            return new ResponseObject<>(null, HttpStatus.BAD_REQUEST,
+                    "Không thể xoá version này vì đang được sử dụng trong sản phẩm!");
+        }
+
         versionRepository.deleteById(id);
         return new ResponseObject<>(null, HttpStatus.OK, "delete thanh cong");
+    }
+
+    @Override
+    public ResponseObject<?> findVersionId(String id){
+        return new ResponseObject<>(versionRepository.findByIdVersion(id)
+                , HttpStatus.OK, "lay chi tiet Version thanh cong");
     }
 }
